refactor: restructure attributes
